+++ server/app/adminapi/logic/shanjian/ShanjianAnchorLogic.php
@@ -2,6 +2,7 @@
 
 namespace app\adminapi\logic\shanjian;
 
+use app\api\logic\DigitalHumanLogic;
 use app\common\logic\BaseLogic;
 use app\common\model\shanjian\ShanjianAnchor;
 
@@ -12,9 +13,25 @@ class ShanjianAnchorLogic extends BaseLogic
     {
         try {
             if (is_string($id)) {
-                ShanjianAnchor::destroy(['id' => $id]);
+                $shanjianAnchor = ShanjianAnchor::findOrEmpty($id);
+                if ($shanjianAnchor->isEmpty()) {
+                    throw new \Exception('数据不存在');
+                }
+
+                if ($shanjianAnchor['dh_id'] !== 0){
+                    DigitalHumanLogic::deletePublicAnchor(['id'=>$shanjianAnchor['dh_id']]);
+                }else{
+                    ShanjianAnchor::destroy(['id' => $id]);
+                }
             } else {
-                ShanjianAnchor:: whereIn('id', $id)   ->select()->delete();
+                $shanjianAnchors = ShanjianAnchor:: whereIn('id', $id)->select();
+                foreach ($shanjianAnchors as $shanjianAnchor) {
+                    if ($shanjianAnchor['dh_id'] !== 0){
+                        DigitalHumanLogic::deletePublicAnchor(['id'=>$shanjianAnchor['dh_id']]);
+                    }else{
+                        ShanjianAnchor::destroy(['id' => $shanjianAnchor['id']]);
+                    }
+                }
             }
             return true;
         } catch (\Exception $e) {
